Added Try, error-checking to initialize
We've observed that in a Nifi workflow, if Accumulo and Zookeepers are not up and running, then the GeoMesaIngestProcessor, if in a RUNNING status, will hang attempting to contact the server.  In this condition, it appears that STOPPING the processor takes about 1-2 minutes.  Stopping and starting Nifi does not remediate the issue.  In a workflow with multiple hung GeoMesaIngestProcessors, it may be faster to delete and rebuild the Nifi directory then attempting to stop them.
The "Try" I have added appears to fix the issue.  The "If" statement with the "ruok" functionality has not been tested in a workflow, but the syntax is correct.  This is the first scala I have written, please amend/discard as appropriate.